+++ docs/changelog.md
@@ -31,6 +31,10 @@
 
 ## 更新说明
 
+1.6.90 fix: fix macro_china_cpi_yearly interface
+
+    1. 修复 macro_china_cpi_yearly 接口，规范输出格式
+
 1.6.89 fix: fix stock_board_concept_name_ths interface
 
     1. 修复 stock_board_concept_name_ths 接口，增加字段说明
@@ -1062,6 +1066,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.90 fix: fix macro_china_cpi_yearly interface
 
 1.6.89 fix: fix stock_board_concept_name_ths interface